test: Try to enable test_server

+++ tests/test_server.py
@@ -13,7 +13,7 @@ import time
 import aiotools
 
 
-if os.environ.get('CI', ''):
+if os.environ.get('CI', '') and sys.version_info < (3, 9, 0):
     pytest.skip(
         'skipped to prevent kill CI agents due to signals on CI environments',
         allow_module_level=True,
